# web/docs/static/skills/relayhub-integration/references/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "RelayHub HTTP API",
    "version": "1.0.0",
    "description": "Shipped MVP routes. JSON bodies are limited to 1 MiB; unknown top-level request fields fail. IDs are opaque. HTTPS production origin. No tenant API is shipped; Go and TypeScript SDKs are available in the repository. Operations endpoints also run on the worker private listener. Standard errors use {error:{code,message}}; unknown routes use 404 not_found and unsupported methods 405 method_not_allowed. CORS is disabled by default; RELAYHUB_ALLOWED_ORIGINS enables exact-origin HTTP CORS and WebSocket origins. Allowed preflights return 204; unsupported methods/headers return 403 forbidden. Actual requests still authenticate. HTTP signature and CORS details: /docs/developer/auth.md."
  },
  "servers": [
    {
      "url": "https://relayhub.dungxbuif.com"
    }
  ],
...
  "paths": {
    "/api/v1/admin/session": {
      "post": {
        "operationId": "createAdminSession",
        "summary": "Create a browser Admin session",
        "description": "Authenticate an Admin email/password account and set the __Host-relayhub_admin cookie. The response includes a session-bound CSRF token for non-safe Admin requests.",
        "tags": [
          "Applications"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Browser Admin session created.",
            "headers": {
              "Set-Cookie": {
                "description": "The __Host-relayhub_admin session cookie.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "csrf_token",
                    "expires_at"
                  ],
                  "properties": {
                    "csrf_token": {
                      "type": "string",
                      "minLength": 43,
                      "maxLength": 43
                    },
                    "expires_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "503": {
            "description": "A required dependency is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteAdminSession",
        "summary": "Revoke the current browser Admin session",
        "description": "Cookie-authenticated requests require the X-RelayHub-CSRF header.",
        "tags": [
          "Applications"
        ],
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-RelayHub-CSRF",
            "in": "header",
            "required": false,
            "description": "Required when authenticating with the Admin session cookie.",
            "schema": {
              "type": "string",
              "minLength": 43,
              "maxLength": 43
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session revoked and cookie expired."
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "503": {
            "description": "A required dependency is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "refreshAdminSession",
        "summary": "Refresh the browser Admin session",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Session refreshed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          }
        }
      }
    },
    "/api/v1/admin/dashboard": {
      "get": {
        "operationId": "getAdminDashboard",
        "summary": "Read bounded cluster dashboard state",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "5m",
                "15m",
                "1h",
                "6h",
                "24h"
              ],
              "default": "15m"
            }
          },
          {
            "name": "step",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "1m",
                "5m",
                "15m",
                "1h"
              ],
              "default": "1m"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rolling metrics, live instances and durable delivery counts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
...
    "/api/v1/admin/events": {
      "get": {
        "operationId": "listAdminEvents",
        "summary": "List events across applications with opaque cursor pagination",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 512
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "source_app_id",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event summaries without payloads.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
    "/api/v1/admin/events/{eventID}": {
      "get": {
        "operationId": "getAdminEvent",
        "summary": "Read an event across applications",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "eventID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event detail.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
    "/api/v1/admin/events/{eventID}/timeline": {
      "get": {
        "operationId": "getAdminEventTimeline",
        "summary": "Read the persisted lifecycle timeline for an event",
        "description": "Returns canonical event detail, delivery summaries, safe callback attempt history and an ordered allowlisted lifecycle. It never parses logs or exposes callback credentials and transport internals.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "eventID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Persisted event lifecycle.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventTimeline"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          }
        }
      }
    },
    "/api/v1/admin/dlq": {
      "get": {
        "operationId": "listAdminDeadLetters",
        "summary": "List dead-letter deliveries without replaying them",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 512
            }
          },
          {
            "name": "source_app_id",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "target_app_id",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "sink",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "stream",
                "callback"
              ]
            }
          },
          {
            "name": "reason",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dead-letter summaries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
    "/api/v1/admin/dlq/replay": {
      "post": {
        "operationId": "replayAdminDeadLetters",
        "summary": "Replay an explicit batch of dead-letter deliveries",
        "description": "Atomically replays 1-100 explicit delivery IDs. Every delivery must currently be dead-lettered. The Idempotency-Key is retained for 24 hours and cannot be rebound to another selection. Cookie authentication also requires X-RelayHub-CSRF.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminReplayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replay result. Idempotent-Replayed is true when returning a stored result.",
            "headers": {
              "Idempotent-Replayed": {
                "schema": {
                  "type": "string",
                  "const": "true"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReplayResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          }
        }
      }
    },
...
    "/api/v1/admin/apps/{appID}": {
      "patch": {
        "operationId": "updateAdminApp",
        "summary": "Update application delivery configuration as an Admin",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "appID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApp"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated application.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/App"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          }
        }
      }
    },
...
    "/api/v2/subscriptions/{subscriptionID}/resume": {
      "post": {
        "operationId": "resumeQueueSubscription",
        "tags": [
          "Queue"
        ],
        "security": [
          {
            "AppApiKey": [],
            "AppSignature": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SubscriptionID"
          }
        ],
        "responses": {
          "200": {
            "description": "Resumed subscription."
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "AdminSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-relayhub_admin",
        "description": "Secure HttpOnly browser Admin session. Non-safe methods also require X-RelayHub-CSRF."
      },
      "AppApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RelayHub-Api-Key",
        "description": "Application API key; must accompany timestamp and signature."
      },
      "AppSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RelayHub-Signature",
        "description": "Lowercase hex HMAC-SHA256 of timestamp + LF + method + LF + exact request target (path and query) + LF + lowercase SHA256 of exact body bytes. No trailing LF. Use the UTF-8 bytes of the entire issued HMAC secret; do not decode its prefix or suffix. See /docs/developer/auth.md."
      },
      "SocketToken": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Short-lived server-issued token with ws:connect or stream:connect scope. URL-encode it; omit query strings from logs."
      }
    },
...
  }
}